requirements.txt
flash_attn==2.7.4.post1; sys_platform == 'linux'
triton-windows ; sys_platform == 'win32'
https://github.com/sdbds/flash-attention-for-windows/releases/download/2.8.3/flash_attn-2.8.3+cu130torch2.9.0cxx11abiFALSEfullbackward-cp311-cp311-win_amd64.whl; sys_platform == 'win32'
librosa>=0.11.0
nnAudio>=0.3.3
torch==2.9.0
torchaudio
torchcodec
torchvision
transformers==4.56.1
wheel>=0.45.1
https://github.com/liyaodev/fairseq/releases/download/v0.12.3.1/fairseq-0.12.3.1-cp311-cp311-win_amd64.whl; sys_platform == 'win32'
https://github.com/liyaodev/fairseq/releases/download/v0.12.3.1/fairseq-0.12.3.1-cp311-cp311-linux_x86_64.whl; sys_platform == 'linux'
hatchling
editables
huggingface_hub[hf_xet]>=0.35.2
ffmpeg-python
accelerate
#vllm==0.8.5
